refactor: remove index file and update import paths for useFilesEncoder - AB-27
