--- src/ui/src/composables/useFilesEncoder/index.ts
@@ -1,1 +1,0 @@
-export * from "./useFilesEncoder";
